+++ FantaRPG/src/Entity.cs
@@ -15,7 +15,7 @@ namespace FantaRPG.src
         public Stats Stats;
         protected bool alive = true;
         public bool Alive { get { return alive; } }
-        public Entity(Texture2D texture, int x, int y, int w, int h)
+        public Entity(Texture2D texture, float x, float y, float w, float h)
         {
             Texture = texture;
             Position = new Vector2(x, y);
